Split Code::Bundle code into its own file
This also moves bundle code from gen_perl, gen_python, and gen_ruby
into the Code class, and refactors the functions to take advantage of
being a member of that class.
