+++ src/generate/code.h
@@ -460,6 +460,12 @@ public:
     // property
     Code& Bundle(GenEnum::PropName prop);
 
+    void BundlePerl(const tt_string_vector& parts);
+    void BundlePython(const tt_string_vector& parts);
+    void BundleRuby(const tt_string_vector& parts);
+
+    void AddPythonImageName(const EmbeddedImage* embed);
+
     // Creates a string using either wxSystemSettings::GetColour(name) or wxColour(r, g, b).
     // Generates wxNullColour if the property is empty.
     Code& ColourCode(GenEnum::PropName prop_name);
