+++ src/generate/gen_perl.cpp
@@ -383,47 +383,6 @@ void BaseCodeGenerator::GeneratePerlClass(PANEL_PAGE panel_type)
 #endif
 }
 
-bool PerlBundleCode(Code& code, GenEnum::PropName prop)
-{
-    auto& description = code.node()->as_string(prop);
-    if (description.empty())
-    {
-        code.Add("wxNullBitmap");
-        return false;
-    }
-
-    tt_view_vector parts(description, BMP_PROP_SEPARATOR, tt::TRIM::both);
-
-    if (parts.size() <= 1 || parts[IndexImage].empty())
-    {
-        code.Add("wxNullBitmap");
-        return false;
-    }
-
-    if (parts[IndexType].contains("Art"))
-    {
-        tt_string art_id(parts[IndexArtID]);
-        tt_string art_client;
-        if (auto pos = art_id.find('|'); tt::is_found(pos))
-        {
-            art_client = art_id.subview(pos + 1);
-            art_id.erase(pos);
-        }
-
-        code << "Wx::ArtProvider::GetBitmap(" << art_id;
-        if (art_client.size())
-            code << ", " << art_client;
-        wxSize art_size { 16, 16 };
-        art_size = GetSizeInfo(parts[IndexSize]);
-
-        // Size is a hack for now...
-        code << ", Wx::Size->new(" << art_size.x << ", " << art_size.y << "))";
-        return true;
-    }
-
-    return false;
-}
-
 tt_string MakePerlPath(Node* node)
 {
     auto [path, has_base_file] = Project.GetOutputPath(node->getForm(), GEN_LANG_PERL);
