+++ src/generate/gen_python.cpp
@@ -465,7 +465,7 @@ bool PythonBitmapList(Code& code, GenEnum::PropName prop)
         {
             if (auto embed = ProjectImages.GetEmbeddedImage(iter); embed)
             {
-                AddPythonImageName(code, embed);
+                code.AddPythonImageName(embed);
                 code += ".Bitmap";
                 needs_comma = true;
                 is_embed_success = true;
@@ -488,194 +488,6 @@ bool PythonBitmapList(Code& code, GenEnum::PropName prop)
     }
     code += " ]\n";
     code.UpdateBreakAt();
-
-    return true;
-}
-
-bool PythonBundleCode(Code& code, GenEnum::PropName prop)
-{
-    auto& description = code.node()->as_string(prop);
-    if (description.empty())
-    {
-        code.Add("wxNullBitmap");
-        return false;
-    }
-
-    tt_view_vector parts(description, BMP_PROP_SEPARATOR, tt::TRIM::both);
-
-    if (parts.size() <= 1 || parts[IndexImage].empty())
-    {
-        code.Add("wxNullBitmap");
-        return false;
-    }
-
-    if (parts[IndexType].contains("Art"))
-    {
-        tt_string art_id(parts[IndexArtID]);
-        tt_string art_client;
-        if (auto pos = art_id.find('|'); tt::is_found(pos))
-        {
-            art_client = art_id.subview(pos + 1);
-            art_id.erase(pos);
-        }
-
-        code.Add("wxArtProvider.GetBitmapBundle(").Add(art_id);
-
-        // Note that current documentation states that the client is required, but the header file says otherwise
-        if (art_client.size())
-            code.Comma().Add(art_client);
-
-        if (parts.size() > IndexSize && parts[IndexSize].size())
-        {
-            wxSize svg_size { -1, -1 };
-            svg_size = GetSizeInfo(parts[IndexSize]);
-
-            if (svg_size != wxDefaultSize)
-            {
-                code.Comma();
-                code.CheckLineLength(sizeof("wx.Size(999, 999)))"));
-                code << "wx.Size(" << svg_size.x << ", " << svg_size.y << ')';
-            }
-        }
-
-        code << ')';
-        return true;
-    }
-
-    auto path = MakePythonPath(code.node());
-
-    if (auto bundle = ProjectImages.GetPropertyImageBundle(description); bundle && bundle->lst_filenames.size())
-    {
-        tt_string name(bundle->lst_filenames[0]);
-        name.make_absolute();
-        if (!name.file_exists())
-        {
-            name = Project.ArtDirectory();
-            name.append_filename(bundle->lst_filenames[0]);
-        }
-        name.make_relative(path);
-        name.backslashestoforward();
-
-        if (description.starts_with("SVG"))
-        {
-            auto embed = ProjectImages.GetEmbeddedImage(parts[IndexImage]);
-            ASSERT(embed);
-            tt_string svg_name;
-            if (embed->form != code.node()->getForm())
-            {
-                svg_name = embed->form->as_string(prop_python_file).filename();
-                svg_name.remove_extension();
-                svg_name << '.' << embed->imgs[0].array_name;
-            }
-            else
-            {
-                svg_name = embed->imgs[0].array_name;
-            }
-            code.insert(0, tt_string("_svg_string_ = zlib.decompress(base64.b64decode(") << svg_name << "))\n");
-            code += "wx.BitmapBundle.FromSVG(_svg_string_";
-            wxSize svg_size { -1, -1 };
-            if (parts[IndexSize].size())
-            {
-                svg_size = GetSizeInfo(parts[IndexSize]);
-            }
-            code.Comma().Str("wx.Size(").itoa(svg_size.x).Comma().itoa(svg_size.y) += "))";
-        }
-
-        else if (description.starts_with("XPM"))
-        {
-            code.CheckLineLength(name.size() + sizeof("wx.Bitmap()") + sizeof("wx.BITMAP_TYPE_XPM)"));
-            code.Str("wx.Bitmap(").QuotedString(name).Comma().Str("wx.BITMAP_TYPE_XPM)");
-        }
-
-        else if (bundle->lst_filenames.size() == 1)
-        {
-            code += "wx.BitmapBundle.FromBitmap(";
-            bool is_embed_success = false;
-
-            if (parts[IndexType].starts_with("Embed"))
-            {
-                if (auto embed = ProjectImages.GetEmbeddedImage(bundle->lst_filenames[0]); embed)
-                {
-                    code.CheckLineLength(embed->imgs[0].array_name.size() + sizeof(".Bitmap)"));
-                    AddPythonImageName(code, embed);
-                    code += ".Bitmap)";
-                    is_embed_success = true;
-                }
-            }
-
-            if (!is_embed_success)
-            {
-                code.CheckLineLength(name.size() + sizeof("wx.Bitmap()"));
-                code.Str("wx.Bitmap(").QuotedString(name) += "))";
-            }
-        }
-        else if (bundle->lst_filenames.size() == 2)
-        {
-            code += "wx.BitmapBundle.FromBitmaps(";
-            bool is_embed_success = false;
-
-            if (parts[IndexType].starts_with("Embed"))
-            {
-                if (auto embed = ProjectImages.GetEmbeddedImage(bundle->lst_filenames[0]); embed)
-                {
-                    code.CheckLineLength(embed->imgs[0].array_name.size() + sizeof(".Bitmap"));
-                    AddPythonImageName(code, embed);
-
-                    code += ".Bitmap";
-
-                    if (auto embed2 = ProjectImages.GetEmbeddedImage(bundle->lst_filenames[1]); embed2)
-                    {
-                        code.Comma().CheckLineLength(embed2->imgs[0].array_name.size() + sizeof(".Bitmap)"));
-                        AddPythonImageName(code, embed2);
-                        code += ".Bitmap)";
-                    }
-                    else
-                    {
-                        code.Comma().Str("wx.NullBitmap)");
-                    }
-                    is_embed_success = true;
-                }
-            }
-            if (!is_embed_success)
-            {
-                tt_string name2(bundle->lst_filenames[1]);
-                name2.make_absolute();
-                name2.make_relative(path);
-                name2.backslashestoforward();
-
-                code.CheckLineLength(name.size() + name2.size() + 27);
-                code.Str("wx.Bitmap(").QuotedString(name).Str(", wx.Bitmap(").QuotedString(name2).Str("))");
-            }
-        }
-        else
-        {
-            code += "wx.BitmapBundle.FromBitmaps([";
-            for (size_t idx = 0; idx < bundle->lst_filenames.size(); ++idx)
-            {
-                if (parts[IndexType].starts_with("Embed"))
-                {
-                    if (auto embed = ProjectImages.GetEmbeddedImage(bundle->lst_filenames[idx]); embed)
-                    {
-                        code.CheckLineLength(embed->imgs[0].array_name.size() + sizeof(".Bitmap"));
-                        AddPythonImageName(code, embed);
-
-                        code += ".Bitmap";
-                        if (idx < bundle->lst_filenames.size() - 1)
-                        {
-                            code.Comma();
-                        }
-                    }
-                }
-            }
-            code += "])";
-        }
-    }
-    else
-    {
-        FAIL_MSG("Missing bundle description");
-        code.Add("wxNullBitmap");
-        return false;
-    }
 
     return true;
 }
