+++ src/generate/gen_ruby.cpp
@@ -610,171 +610,3 @@ tt_string MakeRubyPath(Node* node)
         path.remove_filename();
     return path;
 }
-
-bool RubyBundleCode(Code& code, GenEnum::PropName prop)
-{
-    auto& description = code.node()->as_string(prop);
-    if (description.empty())
-    {
-        code.Add("wxNullBitmap");
-        return false;
-    }
-
-    tt_view_vector parts(description, BMP_PROP_SEPARATOR, tt::TRIM::both);
-
-    if (parts.size() <= 1 || parts[IndexImage].empty())
-    {
-        code.Add("wxNullBitmap");
-        return false;
-    }
-
-    if (parts[IndexType].contains("Art"))
-    {
-        tt_string art_id(parts[IndexArtID]);
-        tt_string art_client;
-        if (auto pos = art_id.find('|'); tt::is_found(pos))
-        {
-            art_client = art_id.subview(pos + 1);
-            art_id.erase(pos);
-        }
-
-        code.Str("Wx::ArtProvider.get_bitmap_bundle(").Add(art_id);
-
-        // Note that current documentation states that the client is required, but the header file says otherwise
-        if (art_client.size())
-            code.Comma().Add(art_client);
-
-        if (parts.size() > IndexSize && parts[IndexSize].size())
-        {
-            wxSize svg_size { -1, -1 };
-            svg_size = GetSizeInfo(parts[IndexSize]);
-
-            if (svg_size != wxDefaultSize)
-            {
-                code.Comma();
-                code.CheckLineLength(sizeof("Wx::Size.new(999, 999)))"));
-                code << "Wx::Size.new(" << svg_size.x << ", " << svg_size.y << ')';
-            }
-        }
-        code << ')';
-        return true;
-    }
-
-    if (auto bundle = ProjectImages.GetPropertyImageBundle(description); bundle && bundle->lst_filenames.size())
-    {
-        if (description.starts_with("SVG"))
-        {
-            auto embed = ProjectImages.GetEmbeddedImage(parts[IndexImage]);
-            ASSERT(embed);
-            tt_string svg_name;
-            if (embed->form != code.node()->getForm())
-            {
-                svg_name = embed->form->as_string(prop_ruby_file).filename();
-                svg_name.remove_extension();
-                svg_name << ".$" << embed->imgs[0].array_name;
-            }
-            else
-            {
-                svg_name = "$" + embed->imgs[0].array_name;
-            }
-            code.insert(0, tt_string("_svg_string_ = Zlib::Inflate.inflate(Base64.decode64(") << svg_name << "))\n");
-            code += "Wx::BitmapBundle.from_svg(_svg_string_";
-            wxSize svg_size { -1, -1 };
-            if (parts[IndexSize].size())
-            {
-                svg_size = GetSizeInfo(parts[IndexSize]);
-            }
-            code.Comma().Str("Wx::Size.new(").itoa(svg_size.x).Comma().itoa(svg_size.y) += "))";
-        }
-
-        else if (description.starts_with("XPM"))
-        {
-            auto path = MakeRubyPath(code.node());
-            tt_string name(bundle->lst_filenames[0]);
-            name.make_absolute();
-            if (!name.file_exists())
-            {
-                name = Project.ArtDirectory();
-                name.append_filename(bundle->lst_filenames[0]);
-            }
-            name.make_relative(path);
-            name.backslashestoforward();
-
-            code.CheckLineLength(name.size() + sizeof("Wx::Bitmap.new()") + sizeof("wx.BITMAP_TYPE_XPM)"));
-            code.Str("Wx::Bitmap.new(").QuotedString(name).Comma().Str("Wx::BITMAP_TYPE_XPM)");
-        }
-
-        else if (parts[IndexType].starts_with("Embed"))
-        {
-            if (bundle->lst_filenames.empty())
-            {
-                code.Add("wxNullBitmap");
-                return false;
-            }
-            if (const EmbeddedImage* embed1 = ProjectImages.GetEmbeddedImage(bundle->lst_filenames[0]); embed1)
-            {
-                code.Str("wxue_get_bundle(").Str("$").Str(embed1->imgs[0].array_name);
-                if (bundle->lst_filenames.size() > 1)
-                {
-                    if (EmbeddedImage* embed2 = ProjectImages.GetEmbeddedImage(bundle->lst_filenames[1]); embed2)
-                    {
-                        code.Comma().Str("$").Str(embed2->imgs[0].array_name);
-                    }
-                    if (bundle->lst_filenames.size() > 2)
-                    {
-                        if (EmbeddedImage* embed3 = ProjectImages.GetEmbeddedImage(bundle->lst_filenames[2]); embed3)
-                        {
-                            code.Comma().Str("$").Str(embed3->imgs[0].array_name);
-                        }
-                    }
-                }
-                code += ')';
-                return true;
-            }
-        }
-        else if (bundle->lst_filenames.size() == 1)
-        {
-            auto path = Project.getBaseDirectory(code.node(), GEN_LANG_RUBY);
-
-            tt_string name(bundle->lst_filenames[0]);
-            name.make_absolute();
-            name.make_relative(path);
-            name.backslashestoforward();
-
-            code.CheckLineLength(name.size() + name.size() + 27);
-            code.Str("Wx::Bitmap(").QuotedString(name).Str("))");
-        }
-        else if (bundle->lst_filenames.size() == 2)
-        {
-            auto path = Project.getBaseDirectory(code.node(), GEN_LANG_RUBY);
-
-            tt_string name(bundle->lst_filenames[0]);
-            name.make_absolute();
-            name.make_relative(path);
-            name.backslashestoforward();
-
-            tt_string name2(bundle->lst_filenames[1]);
-            name2.make_absolute();
-            name2.make_relative(path);
-            name2.backslashestoforward();
-
-            code.CheckLineLength(name.size() + name2.size() + 27);
-            code.Str("Wx::Bitmap(").QuotedString(name).Str(", Wx::Bitmap(").QuotedString(name2).Str("))");
-        }
-
-        else
-        {
-            FAIL_MSG("Unexpected number of images in bundle -- should be <= 2");
-            code.Add("wxNullBitmap");
-            return false;
-        }
-    }
-    else
-    {
-        FAIL_MSG("Missing bundle description");
-        code.Add("wxNullBitmap");
-        return false;
-    }
-
-    return false;
-}
