+++ src/generate/image_gen.h
@@ -11,7 +11,6 @@ class Code;
 struct EmbeddedImage;
 
 std::vector<std::string> base64_encode(unsigned char const* data, size_t data_size, GenLang language = GEN_LANG_PYTHON);
-void AddPythonImageName(Code& code, const EmbeddedImage* embed);
 
 // Generates the code to load a wxBitmapBundle.
 //
